requirements.txt
Django
pyspellchecker
wasmer
spylls
